fixed tests and made requested changes
