+++ src/app/app.component.spec.ts
@@ -75,11 +75,9 @@ import { ThumbnailGridComponent } from './components/thumbnail-grid/thumbnail-gr
 import { NewsFeedComponent } from './components/news-feed/news-feed.component';
 import { MatAutocompleteModule } from '@angular/material';
 import { QueryBarComponent } from './components/query-bar/query-bar.component';
-import {
-    TitleThumbnailSubComponent,
-    DetailsThumbnailSubComponent,
-    CardThumbnailSubComponent
-} from './components/thumbnail-grid/thumbnail-details.component';
+import { DetailsThumbnailSubComponent } from './components/thumbnail-grid/subcomponent.details-view';
+import { TitleThumbnailSubComponent } from './components/thumbnail-grid/subcomponent.title-view';
+import { CardThumbnailSubComponent } from './components/thumbnail-grid/subcomponent.card-view';
 import { TreeModule } from 'angular-tree-component';
 
 describe('App: NeonGtd', () => {
