+++ src/app/components/thumbnail-grid/subcomponent.card-view.ts
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2017 Next Century Corporation
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+import { Component, Input } from '@angular/core';
+import { ThumbnailGridOptions, ThumbnailGridComponent } from './thumbnail-grid.component';
+
+@Component({
+    selector: 'app-subcomponent-card-thumbnail',
+    templateUrl: './subcomponent.card-view.html',
+    styleUrls: ['./thumbnail-grid.component.scss']
+})
+
+export class CardThumbnailSubComponent {
+
+    @Input() item: any;
+    @Input() options: ThumbnailGridOptions;
+
+    thumbnailGrid: ThumbnailGridComponent;
+
+    constructor(grid: ThumbnailGridComponent) {
+        this.thumbnailGrid = grid;
+    }
+}
