+++ src/app/components/visualization-container/visualization-container.component.spec.ts
@@ -51,10 +51,9 @@ import { ThumbnailGridComponent } from '../thumbnail-grid/thumbnail-grid.compone
 import { initializeTestBed } from '../../../testUtils/initializeTestBed';
 import { NewsFeedComponent } from '../news-feed/news-feed.component';
 import { QueryBarComponent } from '../query-bar/query-bar.component';
-import { TitleThumbnailSubComponent,
-    DetailsThumbnailSubComponent,
-    CardThumbnailSubComponent
-} from '../thumbnail-grid/thumbnail-details.component';
+import { DetailsThumbnailSubComponent } from '../thumbnail-grid/subcomponent.details-view';
+import { TitleThumbnailSubComponent } from '../thumbnail-grid/subcomponent.title-view';
+import { CardThumbnailSubComponent } from '../thumbnail-grid/subcomponent.card-view';
 import { MatAutocompleteModule } from '@angular/material';
 import { TreeModule } from 'angular-tree-component';
